VMS-35246: Reorder client files, fix coding style and compiler warnings
Renamed and moved to the corresponding place part of the files, related to screen and voice recording.

Fusion was replaced with `nx_reflect` in part of the functionality.

Some clang compilation warnings are fixed on the way:
* using deprecated datetime locale constants
* overloading which hides virtual functions
* using operator `|` or enums of different types

+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/ui/actions/action_types.h
@@ -118,7 +118,7 @@ enum ActionFlag
     Main = MainScope | NoTarget,
 
     /** Action can appear in scene context menu. */
-    Scene = SceneScope | WidgetTarget,
+    Scene = SceneScope | (int) WidgetTarget,
 
     /**
      * Action can appear in scene context menu in videowall review mode (target elements are
@@ -133,12 +133,12 @@ enum ActionFlag
     Table = TableScope,
 
     /** Action can appear in slider context menu. */
-    Slider = TimelineScope | WidgetTarget,
+    Slider = TimelineScope | (int) WidgetTarget,
 
     /** Action can appear in title bar context menu. */
-    TitleBar = TitleBarScope | LayoutTarget,
+    TitleBar = TitleBarScope | (int) LayoutTarget,
 
-    Notifications = NotificationsScope | WidgetTarget
+    Notifications = NotificationsScope | (int) WidgetTarget
 };
 Q_DECLARE_FLAGS(ActionFlags, ActionFlag)
